+++ lib/ui/explore/ExploreDisplayTypeHeader.dart
@@ -29,8 +29,9 @@ class ExploreDisplayTypeHeader extends StatelessWidget {
   final GestureTapCallback onTapList;
   final GestureTapCallback onTapMap;
   final bool searchVisible;
+  final Map<String, dynamic> additionalData;
 
-  ExploreDisplayTypeHeader({this.displayType, this.onTapList, this.onTapMap, this.searchVisible = false});
+  ExploreDisplayTypeHeader({this.displayType, this.onTapList, this.onTapMap, this.searchVisible = false, this.additionalData});
 
   @override
   Widget build(BuildContext context) {
@@ -72,7 +73,14 @@ class ExploreDisplayTypeHeader extends StatelessWidget {
                       icon: Image.asset('images/icon-search.png'),
                       onPressed: () {
                         Analytics.instance.logSelect(target: "Search");
-                        Navigator.push(context, CupertinoPageRoute(builder: (context) => SearchPanel()));
+                        Navigator.push(context, CupertinoPageRoute(builder: (context) => SearchPanel(searchData:additionalData ))).
+                          then(
+                            (value){
+                              if(value!=null && value == true){
+                                Navigator.pop(context, true);
+                              }
+                            }
+                        );
                       },
                     ),
                   ))